Add ping logs

+++ ping/management/commands/startping.py
@@ -4,6 +4,7 @@ import time
 
 from django.core.management.base import BaseCommand
 
+from ping.models import PingLog
 from visualization.models import Node
 
 DB_REFRESH_RATE = 60
@@ -34,17 +35,21 @@ class Command(BaseCommand):
         return 0
 
     @staticmethod
-    def ping_node(node_ip, limit=100):
+    def ping_node(node_ip):
         node_status = Node.objects.get(ip=node_ip).connected
         while True:
-            p = subprocess.Popen(['ping', '-c', '1', '-W', str(limit), '-i', '0.1', node_ip], stdout=subprocess.PIPE)
-            time.sleep(1)
-            success = p.poll() == 0
-            p.kill()
+            p = subprocess.Popen(['ping', '-c', '1', '-i', '0.2', node_ip], stdout=subprocess.PIPE)
+            time.sleep(0.4)
+            returncode = p.poll()
+            if returncode is None:  # Ping not finished yet,
+                p.kill()
+                returncode = -1
+            success = returncode == 0
             if node_status != success:
                 node_status = success
                 node = Node.objects.get(ip=node_ip)
                 node.connected = success
+                PingLog.objects.create(node=node, connected=success)
                 node.save(update_fields=['connected'])
             if threading.current_thread().stopped():
                 return
